requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aspy.yaml==1.2.0          # via pre-commit
cfgv==1.5.0               # via pre-commit
click==7.0                # via pip-tools
django==2.1.7
djangorestframework-simplejwt==4.1.0
djangorestframework==3.9.2
identify==1.4.0           # via pre-commit
importlib-metadata==0.8   # via pre-commit
nodeenv==1.3.3            # via pre-commit
pip-tools==3.5.0
pre-commit==1.14.4
pudb==2018.1
pygments==2.3.1           # via pudb
pyjwt==1.7.1              # via djangorestframework-simplejwt
pytz==2018.9              # via django
pyyaml==5.1               # via aspy.yaml, pre-commit
six==1.12.0               # via cfgv, pip-tools, pre-commit
toml==0.10.0              # via pre-commit
urwid==2.0.1              # via pudb
virtualenv==16.4.3        # via pre-commit
zipp==0.3.3               # via importlib-metadata
